Classifier log

+++ demos/web/websocket-server.py
@@ -388,13 +388,17 @@ class OpenFaceServerProtocol(WebSocketServerProtocol):
         
         
         os.system('python align-dlib.py'+' '+trainDir+' '+'align outerEyesAndNose Aligned_data/ --size 96')
-        os.remove(alignDir+'/Aligned_data'+'/cache.t7')
+        if os.path.isfile(alignDir+'/Aligned_data'+'/cache.t7'):
+
+            os.remove(alignDir+'/Aligned_data'+'/cache.t7')
         os.system('./batch-represent/main.lua -outDir Feature_gui/ -data Aligned_data/')
         os.system('python classifier.py train Feature_gui/ --classifier RadialSvm')
         
         os.system('cp -r  Unknown/ Train_Image/')
         os.system('python align-dlib.py Train_Image/ align outerEyesAndNose Aligned_data_unknown/ --size 96')
-        os.remove(alignDir+'/Aligned_data_unknown'+'/cache.t7')
+        if os.path.isfile(alignDir+'/Aligned_data_unknown'+'/cache.t7'):
+
+            os.remove(alignDir+'/Aligned_data_unknown'+'/cache.t7')
         os.system('./batch-represent/main.lua -outDir Feature_unknown/ -data Aligned_data_unknown/')
         os.system('python classifier.py train Feature_unknown/ --classifier RadialSvm')
         os.chdir(alignDir+'/Train_Image/')
